requirements.txt
llama-index-core==0.11.11
llama-parse==0.5.6
markdown_pdf==1.3
python-dotenv==1.0.1
langchain-community
sentence_transformers
groq
langchain_groq
